Fix bugs 7 modules + Hoàn thiện màn Quản lý hệ thống thông tin liên lạc VHF

See merge request etc/ttdvcn/bxd.hh/bxd.hh.kcht!139

+++ src/main/java/com/hanghai/kchtg/dikerevetment/dto/DikeRevetmentResponse.java
@@ -10,6 +10,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -33,10 +34,10 @@ public class DikeRevetmentResponse {
     private UUID seaportId;
     private String seaportName;
     private UUID operatingUnitId;
-    private Double length;
-    private Double crestElevation;
+    private BigDecimal length;
+    private BigDecimal crestElevation;
     private LocalDate commissioningDate;
-    private Double height;
+    private BigDecimal height;
     private String surfaceMaterial;
     private String status;
     private String note;
